requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
atomicwrites
attrs
authy
better-profanity
boto3
botocore
certifi
cffi
channels
channels-redis
chardet
colorama
cryptography
dataclasses
deprecation
Django
django-cors-headers
django-storages
djangorestframework
djangorestframework-simplejwt
Faker
frozenlist
# gotrue
gunicorn
h11
h2
hpack
httpcore
httpx>=0.23.0,<0.24.0  # Compatible with postgrest-py 0.10.6
hyperframe
idna
iniconfig
jmespath
msgpack
multidict
ndg-httpsclient
notification
packaging
pillow
pluggy
# postgrest
# postgrest-py>=0.10.6  # Ensure this version is compatible with httpx
psycopg2
pusher
py
pyasn1
pycparser
pydantic
pydantic_core
PyJWT
PyNaCl
pyOpenSSL
pytest
python-dateutil
pytz
realtime
realtime-py
redis
requests
rfc3986
s3transfer
setuptools
six
storage3
StrEnum
# supabase==2.0.1  # Use an older version that might be compatible with httpx<0.24.0
# supabase-py
supafunc
toml
typing_extensions
tzdata
urllib3
websockets
whitenoise
yarl
python-dotenv
